+++ ISH-ADMIN-SERVICE/src/main/java/com/ish/admin/repository/IPlanCategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ish.admin.entity.PlanCategory;
 
-public interface IPlanCategoryRepository extends JpaRepository<PlanCategory, Integer> {
+public interface IPlanCategoryRepository extends JpaRepository<PlanCategory, Long> {
 
 }
