+++ ISH-AR-SERVICE/src/main/java/com/ish/ar/repository/ICitizenAppRegistrationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ish.ar.entity.CitizenApplicationEntity;
 
-public interface ICitizenAppRegistrationRepository extends JpaRepository<CitizenApplicationEntity, Integer> {
+public interface ICitizenAppRegistrationRepository extends JpaRepository<CitizenApplicationEntity, Long> {
 
 }
